@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { DepartmentsService } from '../shared/departments/departments.service';
 
 @Component({
   selector: 'app-categories-menu',
@@ -8,9 +9,14 @@ import { ActivatedRoute } from '@angular/router';
 })
 
 export class SideMenuComponent implements OnInit {
-  constructor(route: ActivatedRoute) {
+  departments: any;
+
+  constructor(route: ActivatedRoute, private departmentsService: DepartmentsService) {
     route.params.subscribe(params => console.log('side menu id parameter', params['id']));
   }
   ngOnInit() {
+    this.departmentsService.getAllDepartments().subscribe((data: any) => {
+      this.departments = data;
+    });
   }
 }
