@@ -1,16 +1,41 @@
-import {Routes} from '@angular/router';
-import {HomeComponent} from './app/home/home.component';
-import {AboutComponent} from './app/about/about.component';
-import {CoursesComponent} from './app/courses/courses.component';
-import {CourseCardsComponent} from './app/course-cards/course-cards.component';
-import {SideMenuComponent} from './app/categories-menu/categories-menu.component';
-import {CoursesCategoryComponent} from './app/course-category/course-category.component';
+import { Routes } from '@angular/router';
+import { HomeComponent } from './app/home/home.component';
+import { AboutComponent } from './app/about/about.component';
+import { CoursesComponent } from './app/courses/courses.component';
+import { CourseCardsComponent } from './app/course-cards/course-cards.component';
+import { SideMenuComponent } from './app/categories-menu/categories-menu.component';
+import { CoursesCategoryComponent } from './app/course-category/course-category.component';
 import { RegisterComponent } from './app/user/register/register.component';
 import { SignInComponent } from './app/user/sign-in/sign-in.component';
+import { SignOutComponent } from './app/user/sign-out/sign-out.component';
 import { AuthGuard } from './app/auth/auth.guard';
 
 export const routerConfig: Routes = [
-    { path: 'home', component: HomeComponent, canActivate: [AuthGuard] },
+    {
+        path: 'home',
+        component: HomeComponent,
+        canActivate: [AuthGuard],
+        children: [
+            {
+                path: '',
+                component: CourseCardsComponent
+            },
+            {
+                path: ':id',
+                component: CoursesCategoryComponent
+            },
+            {
+                path: '',
+                outlet: 'sidemenu',
+                component: SideMenuComponent
+            },
+            {
+                path: ':id',
+                outlet: 'sidemenu',
+                component: SideMenuComponent
+            }
+        ]
+    },
     {
         path: 'about',
         component: AboutComponent
@@ -24,6 +49,10 @@ export const routerConfig: Routes = [
         component: SignInComponent
     },
     {
+        path: 'logoff',
+        component: SignOutComponent
+    },
+    {
         path: 'courses',
         component: CoursesComponent,
         children: [
@@ -32,8 +61,8 @@ export const routerConfig: Routes = [
                 component: CourseCardsComponent
             },
             {
-              path: ':id',
-              component: CoursesCategoryComponent
+                path: ':id',
+                component: CoursesCategoryComponent
             },
             {
                 path: '',
