@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int NumberOfMovies { get; set; }
+        public int NumberOfDoctors { get; set; }
     }
 }
